Crate example_crate

This is a simple (example) Art Library

here lays out the internal structure of the crate, which will then be altered for the public use with use statements

Re-exports

pub use kinds::PrimaryColour;

pub use kinds::SecondaryColour;

pub use utils::mix;

Modules

kinds

utils